requirements.txt
fastapi
uvicorn
passlib==1.7.4
bcrypt==3.2.2
python-dotenv
rich
requests
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
apscheduler==3.10.4
"psycopg[binary]"
PyYAML